fix: move EditProfileForm to components dir (TP-30)
